+++ src/Step/Template/LoadAssets.php
@@ -31,13 +31,13 @@ class LoadAssets implements StepInterface
 
     public function __invoke(Repository $repository, OutputInterface $output)
     {
-        if (! $repository->template) {
+        if (! $repository->metadata['template.directory']) {
             return;
         }
 
         $files = new Finder();
         $files->files()
-            ->in($repository->template->directory)
+            ->in($repository->metadata['template.directory'])
             ->ignoreDotFiles(true)
             ->notName('*.twig');