+++ CHANGELOG.md
@@ -8,6 +8,8 @@ BC breaking changes:
 - the default layout is now `default.twig` (it was previously `page.twig`)
 - template directories are simpler: the `website/public/` directory is not needed anymore, assets should be placed directly in the `website/` directory
 - the config option `templateVariables.baseUrl` is now `baseUrl`
+- the config option `templateUrl` is now `template.url`
+- the config option `directory` is now `template.directory`
 
 Major new features: